floating-icon-app/agent-s2-example/requirements.txt
gui-agents
pyautogui
python-dotenv
orgo
pillow